+++ blosc/trunc-prec.h
@@ -11,7 +11,7 @@
 #define BLOSC_TRUNC_PREC_H
 
 void truncate_precision(uint8_t prec_bits, size_t typesize,
-                        size_t nbytes, const uint8_t* src,
+                        const int32_t nbytes, const uint8_t* src,
                         uint8_t* dest);
 
 #endif //BLOSC_TRUNC_PREC_H